backend/requirements.txt
black==22.10.0
coverage==6.5.0
decorator==5.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.22.1
drf-yasg==1.21.4
lxml==4.9.1
MarkupSafe==2.1.1
nodeenv==1.7.0
numpy==1.24.2
packaging==21.3
Pillow==9.0.1
psutil==5.9.0
ptyprocess==0.7.0
pylev==1.4.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
regex==2023.3.23
requests==2.28.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
setuptools==58.1.0
six==1.16.0
sympy==1.11.1
tokenizers==0.13.2
torch==2.0.0
transformers==4.27.4
typing_extensions==4.5.0
tzdata==2022.6
unittest-xml-reporting==3.2.0
